Fix function props in users page
